feat(api alignment): moving Serializable and Deserializable under their own file.
